Answer and Question Service

+++ src/main/java/dev/astranfalio/teioc/repository/AnswerRepository.java
@@ -4,6 +4,6 @@ import dev.astranfalio.teioc.entity.AnswerEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface AnswerRepository extends JpaRepository<AnswerEntity, Integer> {
+public interface AnswerRepository extends JpaRepository<AnswerEntity, Long> {
 
 }
