+++ src/main/java/dev/astranfalio/teioc/repository/QuestionRepository.java
@@ -4,6 +4,6 @@ import dev.astranfalio.teioc.entity.QuestionEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface QuestionRepository extends JpaRepository<QuestionEntity, Integer> {
+public interface QuestionRepository extends JpaRepository<QuestionEntity, Long> {
 
 }
